requirements.txt
streamlit>=1.32.0
langchain>=0.1.16
langchain-core>=0.1.46
langchain-groq>=0.1.9
python-dotenv>=1.0.1
Pillow>=10.4.0
